Implement public key client authentication.
This commit implements public key client authentication for the Pie CLI.
Before a client can submit an inferlet to the engine, it must authenticate
itself using a public key. The public keys of authorized clients are
managed with the `pie auth` command.

This commit also removes the JWT authentication mechanism.