Refactor the `model` module to use the its own command dispatcher.
The legacy service dispatching mechanism goes through type-erasure
and type downcasting. This is not efficient. Since the model module
is the only module that uses the legacy service dispatching mechanism,
it is better to use its own command dispatcher so that we can avoid the
overhead of type-erasure and type downcasting.

Other modules that used the legacy service dispatching mechanism have
been refactored to use the new `Service` trait.