Fix unintended sleep in scheduler.
When the backend finishes a batched forward pass, the engine will
"unblock" other pending forward pass requests by setting the trigger
variable to true, so that the scheduler can pick them into the next
batch. When there is no available forward pass requests to pick, the
scheduler sleeps.

Previously, there was no notification to the scheduler when the backend
finishes a batched forward pass. The code only sets the trigger variable
to true. The scheduler continues to sleep even if now there are pending
forward pass requests available whose triggers have been set to true.

This commit adds a notification to the scheduler when the backend
finishes a batched forward pass to fix the unintended sleep.